[STUDY-20210210] Components 생성, 반복문 map, 바인딩, 좋아요 버튼
