+++ src/App.css
@@ -1,3 +1,7 @@
+body {
+  background-color: #aaa;
+}
+
 .black-nav {
   background: black;
   width: 100%;
@@ -13,6 +17,14 @@
   text-align: left;
   padding-left: 20px;
   padding-right: 20px;
+}
+
+.listItem {
+  border-bottom: 1px solid pink;
+}
+
+button {
+  display: block;
 }
 
 .modal {
