+++ src/App2.js
